cambio tipo de datos que estaban mal

+++ MercadoEnvio/Clases/ItemFactura.cs
@@ -4,8 +4,8 @@
     {
         public int Id { get; set; }
         public int FacturaId { get; set; }
-        public int Cantidad { get; set; }
-        public double PrecioUnitario { get; set; }
+        public decimal Cantidad { get; set; }
+        public decimal PrecioUnitario { get; set; }
         public string Detalle { get; set; }
     }
 }
